feat: Added fields to charater, and created interface for conditions
Firstly, added an interface Condition for all Conditions which are
applied to a character. Each Condition has a name and description, and
an effect when applied and removed. This is saved in each
Character
Characters have a slice of conditions, as well as new fields like
bulwark, evasion, mainhandattack, and offhand attack which are values
useful for combat. Also added is appropriate getters and setters such as
for conditions